requirements.txt
sqlalchemy==2.0.15
pyyaml==6.0
python-dotenv==1.0.0
pydantic
numpy
multidict
jedi
gunicorn
flask-sqlalchemy==3.0.3
flask-login==0.6.2
flask-httpauth==4.8.0
decorator
aiosignal
aiohttp==3.8.4
werkzeug==2.3.4
jinja2==3.1.2
importlib-metadata
flask==2.2.0
click==8.1.3
urllib3==1.25.8
simplejson==3.16.0
requests==2.22.0
protobuf
pillow
MarkupSafe
certifi
charset-normalizer
docopt
docutils
greenlet
idna==2.8
itsdangerous
pyparsing
python-dateutil
six
typing-extensions
wheel
yarg
zipp
PyJWT==v1.7.1